Replace pipe to bgzip process with direct calls.
The tools using parallel bgzip writing had a occasional loss
by out of synchronization. They are also using CPU resources
efficient enough thanks to huge number of tiles in a flowcell.
So, it is now more efficient to avoid passing output buffers
via interprocess communication, and replace it with direct
calls to bgzf routines in the htslib.

This creates a dependency of development files of htslib.