Trim 3'-side reads for FASTQ output by option.
Late cycles (>= 100th) in 3'-side reads are usually in relatively low
quality in terms of mismatch counts to the reference. Our long set-up
of 3'-side read is primarily for poly(A) length measurements, and they
are virtually useless for sequence-based analyses. Shorter sequence
with high quality would give better result in mappings. In addition,
we can achieve much better efficiency of computation and storage space
by trimming them. This change adds an ability to limit the length of
the 3'-side read, still it can be completely turned off by setting
and option.